split utils into tensorflow utils and general utils
